Big change and cleaning

+++ GameBullet.hpp
@@ -21,7 +21,7 @@ with this program; if not, write to the Free Software Foundation, Inc.,
 
 class GameBullet: public GameObject{
     public:
-        GameBullet(sf::Image &img,int nbrFrame,int nbrLigne,int degat,bool setOnFire,Player *from, int velx,int vely,float height, float width,float offsetColX,float offsetColY,bool col);
+        GameBullet(sf::Image &img,int nbrFrame,int nbrLigne,int degat,bool setOnFire,Player *from, float velx,float vely,float height, float width,float offsetColX,float offsetColY,bool col);
         virtual bool isDraw();
         virtual bool isDelete()const;
         virtual bool isCollision()const;
